refactor: move register_operator function to utils
