+++ operator/rust/crates/operator/src/start_operator.rs
@@ -208,110 +208,6 @@ async fn monitor_new_tasks_polling(rpc_url: &str, private_key: &str) -> Result<(
     }
 }
 
-pub async fn register_operator(rpc_url: &str, private_key: &str) -> Result<()> {
-    let pr = get_signer(private_key, rpc_url);
-    let signer = PrivateKeySigner::from_str(private_key)?;
-
-    let el_data = get_anvil_eigenlayer_deployment_data()?;
-    let delegation_manager_address: Address = el_data.addresses.delegation_manager.parse()?;
-    let avs_directory_address: Address = el_data.addresses.avs_directory.parse()?;
-
-    let elcontracts_reader_instance = ELChainReader::new(
-        get_logger().clone(),
-        None,
-        delegation_manager_address,
-        Address::ZERO,
-        avs_directory_address,
-        None,
-        rpc_url.to_string(),
-    );
-    let elcontracts_writer_instance = ELChainWriter::new(
-        Address::ZERO,
-        Address::ZERO,
-        None,
-        None,
-        Address::ZERO,
-        elcontracts_reader_instance.clone(),
-        rpc_url.to_string(),
-        private_key.to_string(),
-    );
-
-    let operator = Operator {
-        address: signer.address(),
-        delegation_approver_address: Address::ZERO,
-        staker_opt_out_window_blocks: Some(0),
-        metadata_url: Default::default(),
-        allocation_delay: Some(0),
-        _deprecated_earnings_receiver_address: None,
-    };
-
-    let is_registered = elcontracts_reader_instance
-        .is_operator_registered(signer.address())
-        .await
-        .unwrap();
-    get_logger().info(&format!("is registered {}", is_registered), "");
-    let tx_hash = elcontracts_writer_instance
-        .register_as_operator(operator)
-        .await?;
-    let receipt = pr.get_transaction_receipt(tx_hash).await?;
-    if !receipt.is_some_and(|r| r.inner.is_success()) {
-        get_logger().error("Operator registration failed", "");
-        return Err(eyre::eyre!("Operator registration failed"));
-    }
-    get_logger().info(
-        &format!(
-            "Operator registered on EL successfully tx_hash {:?}",
-            tx_hash
-        ),
-        "",
-    );
-    let mut salt = [0u8; 32];
-    rand::rngs::OsRng.try_fill_bytes(&mut salt).unwrap();
-
-    let salt = FixedBytes::from_slice(&salt);
-    let now = Utc::now().timestamp();
-    let expiry: U256 = U256::from(now + 3600);
-
-    let hello_world_contract_address: Address = get_hello_world_service_manager()?;
-    let digest_hash = elcontracts_reader_instance
-        .calculate_operator_avs_registration_digest_hash(
-            signer.address(),
-            hello_world_contract_address,
-            salt,
-            expiry,
-        )
-        .await?;
-
-    let signature = signer.sign_hash_sync(&digest_hash)?;
-    let operator_signature = SignatureWithSaltAndExpiry {
-        signature: signature.as_bytes().into(),
-        salt,
-        expiry,
-    };
-    let stake_registry_address = get_stake_registry_address()?;
-    let contract_ecdsa_stake_registry = ECDSAStakeRegistry::new(stake_registry_address, &pr);
-    let registeroperator_details_call = contract_ecdsa_stake_registry
-        .registerOperatorWithSignature(operator_signature, signer.clone().address())
-        .gas(500000);
-    let register_hello_world_hash = registeroperator_details_call
-        .send()
-        .await?
-        .get_receipt()
-        .await?
-        .transaction_hash;
-
-    get_logger().info(
-        &format!(
-            "Operator registered on AVS successfully :{} , tx_hash :{}",
-            signer.address(),
-            register_hello_world_hash
-        ),
-        "",
-    );
-
-    Ok(())
-}
-
 #[tokio::main]
 pub async fn main() {
     use tokio::signal;
