+++ operator/rust/crates/utils/Cargo.toml
@@ -10,6 +10,9 @@ license-file.workspace = true
 
 [dependencies]
 alloy.workspace = true
+eigensdk.workspace = true
 serde.workspace = true
 serde_json = "1.0.121"
 eyre = "0.6.12"
+rand = "0.9"
+chrono = "0.4.38"
